Remove graphite integration
