+++ helm/upp-aggregate-healthcheck/templates/deployment.yaml
@@ -45,11 +45,6 @@ spec:
             configMapKeyRef:
               name: global-config
               key: cluster.url
-        - name: GRAPHITE_URL
-          valueFrom:
-            configMapKeyRef:
-              name: global-config
-              key: graphite.address
         - name: PATH_PREFIX
           value: {{ .Values.env.PATH_PREFIX }} 
         ports:
