+++ model.go
@@ -3,6 +3,8 @@ package main
 import (
 	"sync"
 	"time"
+
+	fthealth "github.com/Financial-Times/go-fthealth/v1_1"
 )
 
 type pod struct {
@@ -37,9 +39,12 @@ type servicesMap struct {
 	m map[string]service
 }
 
+type bufferedHealths struct {
+	buffer chan fthealth.CheckResult
+}
+
 type measuredService struct {
 	service         service
 	cachedHealth    *cachedHealth
-	bufferedHealths *bufferedHealths
 	bufferedMetrics *bufferedHealths
 }
